+++ CHANGELOG.md
@@ -12,6 +12,12 @@ This project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.
 
 ### Added
 
+- Functions `setup_laplace_matrix()` and `setup_mass_matrix()` compute those
+  two matrices for both triangle meshes and general polygon meshes.
+  The general case is based on the paper Bunge et al, "Polygon Laplacian made simple",
+  Eurographics 2020.
+- Smoothing, parameterization, and fairing are now implemented based on
+  sparse Laplacian matrices, which generalizes to general polygon meshes.
 - Add `PMP_STRICT_COMPILATION` CMake option to control treating warnings as errors. Default: On.
 - Add function `flip_faces()` to reverse face orientation in a mesh (#123)
 - Meshes and textures can be loaded by dropping the files onto the window, both in
@@ -20,6 +26,13 @@ This project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.
 ### Changed
 
 - `SurfaceMeshGL` renamed to `Renderer` and de-coupled from `SurfaceMesh`
+- Replace `triangle_area(Face)` by `face_area(Face)`, which now
+  supports general polygons. `surface_area(SurfaceMesh)` now also
+  works for general polygon meshes.
+- Replace mixed Voronoi area by barycentric Voronoi area in
+  `voronoi_area()`, because it generalizes better to polygon meshes.
+  Remove `voronoi_area_barycentric()`, because it's now computed by
+  `voronoi_area()`
 - Remove max angle triangulation objective leading to fold-overs.
 - Breaking change: Re-design algorithms API to use a simple function interface.
 - Breaking change: Make helper classes `Heap`, `Quadric`, `NormalCone`, `TriangleKdTree` private.